+++ Blurred-ios/Channels/Video/Overlays/Side Overlay/OverlayViewDelegate.swift
@@ -11,5 +11,5 @@ import Foundation
 protocol ChannelVideoOverlayViewDelegate: AnyObject {
     func didTapChannel(_ view:ChannelVideoOverlayView, videousername: String)
     func didTapComments(_ view:ChannelVideoOverlayView, videoid: Int)
-    func didTapShare(_ view:ChannelVideoOverlayView, videoUrl: String)
+    func didTapShare(_ view:ChannelVideoOverlayView, videoUrl: String, videoId: Int)
 }